Fixed Random HDD Seek throughput
Hi! Thanks for this list!

Maybe I am missing something trivial, but I think Random HDD read throughput (assuming repeated random seek + 8KB read) is closer to 0.7 than 70 MB/s. Fixing it if it was a typo. If not, happy to learn where I am wrong (: 

Is it because we assume here Seek + Sequential read throughput?